+++ src/main/java/roomescape/controller/api/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import jakarta.validation.Valid;
 import java.net.URI;
